feat: add search endpoint - move funcs in line with GoKoala
